Split javascript into notifications handling and content action logging parts
This change separates the Nitro library initialization from including
the local javascript. The local part was split up into notifications
handling and content action logging parts. This way it is easy to
control where to include what javascript. Notification handling is
included on all page requests, so that e.g. user profile related
notifications are shown in the proper place.